menu: Automatically scroll down the menu if up/down/left/right is held
Allows a user of the mod to scroll up and down menus which is much more
convenient than having to continually press up or down repeatedly to
reach the desired menu selection.

We can also handle left and right dpad presses to make it convenient
when editing amounts as well.

We currently cap the speed at 75 ms as this allows the menu to scroll at
a reasonable speed with a delay long enough that the user won't
accidentally skip over what they were looking for.